Modify /Gemfile and /init.rb, move /lib to /app.
