+++ Gemfile
@@ -2,6 +2,11 @@
 
 source 'https://rubygems.org'
 
+# Web Application
+gem 'puma', '~> 5.5'
+gem 'roda', '~> 3.49'
+gem 'slim', '~> 4.1'
+
 # Validation
 gem 'dry-struct', '~> 1.4'
 gem 'dry-types', '~> 1.5'
@@ -12,6 +17,7 @@ gem 'http', '~> 5.0'
 # Testing
 gem 'minitest', '~> 5.0'
 gem 'minitest-rg', '~> 5.0'
+gem 'rerun', '~> 0'
 gem 'simplecov', '~> 0'
 gem 'vcr', '~> 6.0'
 gem 'webmock', '~> 3.0'
