+++ app/lib/entities/feed.rb
