+++ app/lib/init.rb
