+++ app/lib/project_info.rb
