+++ config.ru
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative './init.rb'
+run Headline_Connector::App.freeze.app
